Revert "Merge warnings."
This reverts commit c9f91a1238933ae3891ae7706df600ddcf352ef7.
